+++ Geometry3D/geometry/__init__.py
@@ -3,15 +3,19 @@ from .line import Line,x_axis,y_axis,z_axis
 from .plane import Plane,xy_plane,yz_plane,xz_plane
 from .point import Point,origin
 from .segment import Segment
-from .polygen import ConvexPolygen,Parallelogram
+from .polygen import ConvexPolygen,Parallelogram,get_circle_point_list,Circle
 from .pyramid import Pyramid
-from .polyhedron import ConvexPolyhedron,Parallelepiped
+from .polyhedron import ConvexPolyhedron,Parallelepiped,Sphere,Cylinder,Cone
 from .halfline import HalfLine
 __all__ = (
     "ConvexPolyhedron",
     "Parallelepiped",
+    "Sphere",
+    "Cone",
+    "Cylinder",
     "ConvexPolygen",
     "Parallelogram",
+    "Circle",
     "Pyramid",
     "Segment",
     "Line",
@@ -25,4 +29,5 @@ __all__ = (
     "xy_plane",
     "yz_plane",
     "xz_plane",
+    "get_circle_point_list"
 )