+++ Geometry3D/utils/__init__.py
@@ -1,6 +1,6 @@
 from .solver import solve
 from .vector import Vector,x_unit_vector,y_unit_vector,z_unit_vector
-from .constant import set_eps,get_eps,get_sig_figures,set_sig_figures
+from .constant import set_eps,get_eps,get_sig_figures,set_sig_figures,SMALL_ANGLE
 from .logger import set_log_level,get_main_logger
 __all__=(
     "solve",
@@ -10,6 +10,7 @@ __all__=(
     "z_unit_vector",
     "set_eps",
     "get_eps",
+    "SMALL_ANGLE",
     "get_sig_figures",
     "set_sig_figures",
     "set_log_level",