+++ changelog.md
@@ -1,4 +1,5 @@
 # 0.2.0
 - add HalfLine geometry
 - modify `__repr__` for ConvexPolygen
-- add sphere, cone, cylinder
+- add Sphere, Cylinder and Cone class method for ConvexPolyhedron
+- add Circle classmethod for Convexpolygen